fix sth

+++ laravel/database/migrations/2021_07_20_090747_create_users_table.php
@@ -15,8 +15,8 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->integer("id")->autoIncrement();
-            $table->integer("facebook_id")->nullable();
-            $table->integer("google_id")->nullable();
+            $table->bigInteger("facebook_id")->nullable();
+            $table->bigInteger("google_id")->nullable();
             $table->string("email")->nullable();
             $table->string("name")->nullable();
             $table->string("phone")->nullable();
